test: 각 클래스들로 테스트들을 분류
각 클래스들의 기능들의 테스트들을 클래스별로 분류하였습니다.
